backend/requirements.txt
fastapi[all]
uvicorn
PyGithub
pandas
nltk
python-dotenv
google-generativeai